pyproject.toml
[project]
name = "ainewssprite"
version = "0.1.0"
description = "AI 新闻精灵 -- 每日 AI 领域新闻自动聚合工具"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0.10",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "python-dateutil>=2.8.2",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.5.0",
]

[project.scripts]
ainewssprite = "ainewssprite.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
